server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
openai>=1.84.0
python-multipart>=0.0.20
python-dotenv>=1.1.0

pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.26.0
httpx>=0.24.0

ruff>=0.1.0
isort>=5.10.0
mypy>=1.0.0
pre-commit>=3.0.0
